rest: Remove controller advice that is only applicable one controller
