+++ budgeteer-rest/src/main/java/de/adesso/budgeteer/rest/project/ProjectController.java
@@ -2,6 +2,8 @@ package de.adesso.budgeteer.rest.project;
 
 import de.adesso.budgeteer.core.project.ProjectException;
 import de.adesso.budgeteer.core.project.port.in.*;
+import de.adesso.budgeteer.rest.project.errors.CreateProjectError;
+import de.adesso.budgeteer.rest.project.errors.UpdateProjectError;
 import de.adesso.budgeteer.rest.project.exceptions.CreateProjectException;
 import de.adesso.budgeteer.rest.project.exceptions.UpdateProjectException;
 import de.adesso.budgeteer.rest.project.model.*;
@@ -13,6 +15,7 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
@@ -128,5 +131,31 @@ public class ProjectController {
       @PathVariable("projectId") ProjectIdModel projectIdModel) {
     return projectModelMapper.toModel(
         getProjectWithDateUseCase.getProjectWithDate(projectIdModel.getValue()));
+  }
+
+  @ExceptionHandler(CreateProjectException.class)
+  public ResponseEntity<CreateProjectError> handleCreateProjectException(
+      CreateProjectException createProjectException) {
+    var causes = createProjectException.getProjectException().getCauses().getAll();
+
+    var createProjectError = new CreateProjectError();
+    createProjectError.setNamealreadyInUse(
+        causes.contains(ProjectException.ProjectErrors.NAME_ALREADY_IN_USE));
+
+    return ResponseEntity.badRequest().body(createProjectError);
+  }
+
+  @ExceptionHandler(UpdateProjectException.class)
+  public ResponseEntity<UpdateProjectError> handleUpdateProjectException(
+      UpdateProjectException updateProjectException) {
+    var causes = updateProjectException.getProjectException().getCauses().getAll();
+
+    var updateProjectError = new UpdateProjectError();
+    updateProjectError.setProjectNotFound(
+        causes.contains(ProjectException.ProjectErrors.PROJECT_NOT_FOUND));
+    updateProjectError.setNameAlreadyInUse(
+        causes.contains(ProjectException.ProjectErrors.NAME_ALREADY_IN_USE));
+
+    return ResponseEntity.badRequest().body(updateProjectError);
   }
 }
